refactor chat merge util fie
