+++ whatstk/objects.py
@@ -1,7 +1,7 @@
 from whatstk.utils.parser import generate_regex, parse_chat, remove_alerts_from_df
 from whatstk.utils.auto_header import extract_header_from_text
 from whatstk.utils.exceptions import RegexError, HFormatError
-from whatstk.analysis.process import merge_chats
+from whatstk.utils.chat_merge import merge_chats
 
 
 class WhatsAppChat:
